webservice/requirements.txt
bottle
librosa
boto3
pynamodb
python-dotenv
requests